# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/schedules": {
      "get": {
        "description": "Returns all schedules, optionally filtered by scheduleIds.",
        "operationId": "ScheduleController_findAll",
        "parameters": [
          {
            "name": "scheduleIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of schedule IDs to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all schedules",
        "tags": [
          "Schedule"
        ]
      },
      "post": {
        "operationId": "ScheduleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/api/schedules/search": {
      "get": {
        "description": "Search for schedules by pools, instructors, days of week, date, timezone, and includeReserved.",
        "operationId": "ScheduleController_search",
        "parameters": [
          {
            "name": "pools",
            "required": false,
            "in": "query",
            "description": "List of pool IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "instructors",
            "required": false,
            "in": "query",
            "description": "List of instructor IDs",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "daysOfWeek",
            "required": false,
            "in": "query",
            "description": "List of days of the week",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Date to search for (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timezone",
            "required": false,
            "in": "query",
            "description": "Timezone",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeReserved",
            "required": false,
            "in": "query",
            "description": "Include reserved schedules",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchScheduleResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Search schedules",
        "tags": [
          "Schedule"
        ]
      }
    },
    "/api/schedules/available-dates": {
      "get": {
        "operationId": "ScheduleController_findAvailableDates",
        "parameters": [
          {
            "name": "timezone",
            "required": false,
            "in": "query",
            "description": "IANA timezone (e.g., America/Denver)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      }
    },
    "/api/schedules/{id}": {
      "get": {
        "operationId": "ScheduleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      },
      "patch": {
        "operationId": "ScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      },
      "delete": {
        "operationId": "ScheduleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Schedule"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stansbury Swim API",
    "description": "API documentation for Stansbury Swim backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ScheduleStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "birthDate",
          "notes"
        ]
      },
      "ScheduleUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "instructor"
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "role"
        ]
      },
      "ScheduleRegistrationDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "pending": {
            "type": "boolean",
            "description": "True when the seat is held pending payment."
          },
          "student": {
            "$ref": "#/components/schemas/ScheduleStudentDto"
          },
          "user": {
            "$ref": "#/components/schemas/ScheduleUserDto"
          }
        },
        "required": [
          "studentId",
          "userId",
          "createdAt",
          "student",
          "user"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "poolId": {
            "type": "string"
          },
          "instructorId": {
            "type": "string"
          },
          "classSize": {
            "type": "number"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "startDateTime": {
            "type": "string"
          },
          "endDateTime": {
            "type": "string"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleRegistrationDto"
            }
          }
        },
        "required": [
          "id",
          "poolId",
          "instructorId",
          "classSize",
          "lessonType",
          "startDateTime",
          "endDateTime",
          "registrations"
        ]
      },
...
      "SearchScheduleStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "birthDate",
          "notes"
        ]
      },
      "SearchScheduleUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin",
              "instructor"
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "role"
        ]
      },
      "SearchScheduleRegistrationDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "pending": {
            "type": "boolean",
            "description": "True when the seat is held pending payment."
          },
          "student": {
            "$ref": "#/components/schemas/SearchScheduleStudentDto"
          },
          "user": {
            "$ref": "#/components/schemas/SearchScheduleUserDto"
          }
        },
        "required": [
          "studentId",
          "userId",
          "createdAt",
          "student",
          "user"
        ]
      },
      "SearchScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "poolId": {
            "type": "string"
          },
          "instructorId": {
            "type": "string"
          },
          "classSize": {
            "type": "number"
          },
          "lessonType": {
            "type": "string",
            "enum": [
              "private",
              "group"
            ]
          },
          "startDateTime": {
            "type": "string"
          },
          "endDateTime": {
            "type": "string"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchScheduleRegistrationDto"
            }
          }
        },
        "required": [
          "id",
          "poolId",
          "instructorId",
          "classSize",
          "lessonType",
          "startDateTime",
          "endDateTime",
          "registrations"
        ]
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}